Add files via upload
Review of recent (as of March 2019) applications of the α-relative Rényi entropies to the AdS/CFT duality, holography, and the c-theorem. Presented at the II Workshop on Quantum Information and Thermodynamics on 2019 March 11 at the International Institute of Physics at the Federal University of Rio Grande do Norte, Brazil.